+++ cmd/sqlpilot/main.go
@@ -1,4 +1,4 @@
-﻿package main
+package main
 
 import (
 	"fmt"
@@ -10,7 +10,7 @@ import (
 )
 
 var (
-	version = "0.2.0"
+	version = "0.3.0"
 )
 
 func main() {